Abstract class implemented (GridOfButtons).
GridOfButtons class created to centralise functions that are core to
grids filled with buttons (NetworkQueue/EditingGrid).
These grids are now dynamically created in python as opposed to being
explicitly created in KV.
GridHeaders and the Send/Log buttons in the Networkqueue are now
seperate from the grid itself. (This is much nicer to deal with and
allows for the GridOfButtons class to work).