Column sizes
The columns and their corresponding labels can have their
width/size_hint_x set now.
This means that the ID column, is appropriately short, as it can only
contain 4 characters.
Their size_hint is controlled by a x_hint_list in both of the respective
Grid classes. (This also means that the labels are added in python as
opposed to KV).