Small but potent touch.
on_touch_down has a dramatically different effect to on_touch_up with
the scrollview. on_touch_up would trigger twice, and would be triggered
outside the bounding box of the scrollview.
I think it's due to the scrollview grabbing the touch on 'on_touch_down'
and then, well it has access to that touch on 'on_touch_up'