Split Classes, moved functions, Auto-fill.
TextInputs now automatically fill with the appropriate data from the
file (easier to edit).
UserInput was split into two classes, as it did too many things and was
becoming rather large.

info_from_file was changed from a method to be a function in parse_kivy
as there is no reason for it to be specific to the EditingGrid class.

Class attributes were added to work with info_from-file being in
parse_kivy.